+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Fixes
 - `npcrange` objective is triggered at wrong time
+- Citizens compatibility for not spawned NPCs
 ### Security
 
 ## [1.12.1] - 2021-02-05